Support ManagementFactory.getPlatformMBeanServer
