+++ substratevm/src/com.oracle.svm.core.genscavenge/src/com/oracle/svm/core/genscavenge/HeapImpl.java
@@ -25,19 +25,11 @@
 package com.oracle.svm.core.genscavenge;
 
 //Checkstyle: stop
-import java.lang.management.ManagementFactory;
-import java.lang.management.MemoryMXBean;
-import java.lang.management.MemoryUsage;
+
 import java.lang.ref.Reference;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
-
-import javax.management.MBeanNotificationInfo;
-import javax.management.NotificationEmitter;
-import javax.management.NotificationFilter;
-import javax.management.NotificationListener;
-import javax.management.ObjectName;
 
 import org.graalvm.compiler.api.replacements.Fold;
 import org.graalvm.compiler.nodes.gc.CardTableBarrierSet;
@@ -47,14 +39,9 @@ import org.graalvm.nativeimage.Platform;
 import org.graalvm.nativeimage.Platforms;
 import org.graalvm.nativeimage.hosted.Feature.FeatureAccess;
 import org.graalvm.word.Pointer;
-import org.graalvm.word.PointerBase;
 import org.graalvm.word.UnsignedWord;
-import org.graalvm.word.WordFactory;
 
 import com.oracle.svm.core.MemoryWalker;
-import com.oracle.svm.core.MemoryWalker.CodeAccess;
-import com.oracle.svm.core.MemoryWalker.HeapChunkAccess;
-import com.oracle.svm.core.MemoryWalker.NativeImageHeapRegionAccess;
 import com.oracle.svm.core.SubstrateOptions;
 import com.oracle.svm.core.SubstrateUtil;
 import com.oracle.svm.core.annotate.AlwaysInline;
@@ -62,7 +49,6 @@ import com.oracle.svm.core.annotate.NeverInline;
 import com.oracle.svm.core.annotate.Substitute;
 import com.oracle.svm.core.annotate.TargetClass;
 import com.oracle.svm.core.annotate.Uninterruptible;
-import com.oracle.svm.core.code.CodeInfo;
 import com.oracle.svm.core.heap.GC;
 import com.oracle.svm.core.heap.GCCause;
 import com.oracle.svm.core.heap.Heap;
@@ -78,7 +64,6 @@ import com.oracle.svm.core.locks.VMMutex;
 import com.oracle.svm.core.log.Log;
 import com.oracle.svm.core.nodes.CFunctionEpilogueNode;
 import com.oracle.svm.core.nodes.CFunctionPrologueNode;
-import com.oracle.svm.core.option.RuntimeOptionValues;
 import com.oracle.svm.core.snippets.KnownIntrinsics;
 import com.oracle.svm.core.thread.JavaThreads;
 import com.oracle.svm.core.thread.ThreadStatus;
@@ -87,7 +72,6 @@ import com.oracle.svm.core.thread.VMThreads;
 
 import jdk.vm.ci.meta.MetaAccessProvider;
 import jdk.vm.ci.meta.ResolvedJavaType;
-import sun.management.Util;
 //Checkstyle: resume
 
 /** An implementation of a card remembered set generational heap. */
@@ -103,7 +87,6 @@ public final class HeapImpl extends Heap {
     private final ObjectHeaderImpl objectHeaderImpl = new ObjectHeaderImpl();
     private final GCImpl gcImpl;
     private final HeapPolicy heapPolicy;
-    private final MemoryMXBean memoryMXBean = new HeapImplMemoryMXBean();
     private final ImageHeapInfo imageHeapInfo = new ImageHeapInfo();
     private HeapVerifier heapVerifier;
     private final StackVerifier stackVerifier;
@@ -385,11 +368,6 @@ public final class HeapImpl extends Heap {
             // @formatter:on
         }
         return log;
-    }
-
-    @Override
-    public MemoryMXBean getMemoryMXBean() {
-        return memoryMXBean;
     }
 
     /** Return a list of all the classes in the heap. */
@@ -736,146 +714,6 @@ public final class HeapImpl extends Heap {
         } finally {
             REF_MUTEX.unlock();
         }
-    }
-}
-
-/**
- * A MemoryMXBean for this heap.
- *
- * Note: This implementation is somewhat inefficient, in that each time it is asked for the
- * <em>current</em> heap memory usage or non-heap memory usage, it uses the MemoryWalker.Visitor to
- * walk all of memory. If someone asks for only the heap memory usage <em>or</em> the non-heap
- * memory usage, the other kind of memory will still be walked. If someone asks for both the heap
- * memory usage <em>and</em> the non-heap memory usage, all the memory will be walked twice.
- */
-final class HeapImplMemoryMXBean implements MemoryMXBean, NotificationEmitter {
-    static final long UNDEFINED_MEMORY_USAGE = -1L;
-
-    private final MemoryMXBeanMemoryVisitor visitor;
-
-    @Platforms(Platform.HOSTED_ONLY.class)
-    HeapImplMemoryMXBean() {
-        this.visitor = new MemoryMXBeanMemoryVisitor();
-    }
-
-    @Override
-    public ObjectName getObjectName() {
-        return Util.newObjectName(ManagementFactory.MEMORY_MXBEAN_NAME);
-    }
-
-    @Override
-    public int getObjectPendingFinalizationCount() {
-        return 0;
-    }
-
-    @Override
-    public MemoryUsage getHeapMemoryUsage() {
-        visitor.reset();
-        MemoryWalker.getMemoryWalker().visitMemory(visitor);
-        long used = visitor.getHeapUsed().rawValue();
-        long committed = visitor.getHeapCommitted().rawValue();
-        return new MemoryUsage(UNDEFINED_MEMORY_USAGE, used, committed, UNDEFINED_MEMORY_USAGE);
-    }
-
-    @Override
-    public MemoryUsage getNonHeapMemoryUsage() {
-        visitor.reset();
-        MemoryWalker.getMemoryWalker().visitMemory(visitor);
-        long used = visitor.getNonHeapUsed().rawValue();
-        long committed = visitor.getNonHeapCommitted().rawValue();
-        return new MemoryUsage(UNDEFINED_MEMORY_USAGE, used, committed, UNDEFINED_MEMORY_USAGE);
-    }
-
-    @Override
-    public boolean isVerbose() {
-        return SubstrateOptions.PrintGC.getValue();
-    }
-
-    @Override
-    public void setVerbose(boolean value) {
-        RuntimeOptionValues.singleton().update(SubstrateOptions.PrintGC, value);
-    }
-
-    @Override
-    public void gc() {
-        System.gc();
-    }
-
-    @Override
-    public void removeNotificationListener(NotificationListener listener, NotificationFilter filter, Object handback) {
-    }
-
-    @Override
-    public void addNotificationListener(NotificationListener listener, NotificationFilter filter, Object handback) {
-    }
-
-    @Override
-    public void removeNotificationListener(NotificationListener listener) {
-    }
-
-    @Override
-    public MBeanNotificationInfo[] getNotificationInfo() {
-        return new MBeanNotificationInfo[0];
-    }
-}
-
-/** A MemoryWalker.Visitor that records used and committed memory sizes. */
-final class MemoryMXBeanMemoryVisitor implements MemoryWalker.Visitor {
-    private UnsignedWord heapUsed;
-    private UnsignedWord heapCommitted;
-    private UnsignedWord nonHeapUsed;
-    private UnsignedWord nonHeapCommitted;
-
-    MemoryMXBeanMemoryVisitor() {
-        reset();
-    }
-
-    public UnsignedWord getHeapUsed() {
-        return heapUsed;
-    }
-
-    public UnsignedWord getHeapCommitted() {
-        return heapCommitted;
-    }
-
-    public UnsignedWord getNonHeapUsed() {
-        return nonHeapUsed;
-    }
-
-    public UnsignedWord getNonHeapCommitted() {
-        return nonHeapCommitted;
-    }
-
-    public void reset() {
-        heapUsed = WordFactory.zero();
-        heapCommitted = WordFactory.zero();
-        nonHeapUsed = WordFactory.zero();
-        nonHeapCommitted = WordFactory.zero();
-    }
-
-    @Override
-    public <T> boolean visitNativeImageHeapRegion(T region, NativeImageHeapRegionAccess<T> access) {
-        UnsignedWord size = access.getSize(region);
-        heapUsed = heapUsed.add(size);
-        heapCommitted = heapCommitted.add(size);
-        return true;
-    }
-
-    @Override
-    public <T extends PointerBase> boolean visitHeapChunk(T heapChunk, HeapChunkAccess<T> access) {
-        UnsignedWord used = access.getAllocationEnd(heapChunk).subtract(access.getAllocationStart(heapChunk));
-        UnsignedWord committed = access.getSize(heapChunk);
-        heapUsed = heapUsed.add(used);
-        heapCommitted = heapCommitted.add(committed);
-        return true;
-    }
-
-    @Override
-    public <T extends CodeInfo> boolean visitCode(T codeInfo, CodeAccess<T> access) {
-        UnsignedWord size = access.getSize(codeInfo).add(access.getMetadataSize(codeInfo));
-        nonHeapUsed = nonHeapUsed.add(size);
-        nonHeapCommitted = nonHeapCommitted.add(size);
-        return true;
     }
 }
