+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/image/DisallowedImageHeapObjectFeature.java
@@ -24,13 +24,17 @@
  */
 package com.oracle.svm.hosted.image;
 
+import java.lang.management.PlatformManagedObject;
 import java.nio.charset.Charset;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.IdentityHashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
+
+import javax.management.MBeanServerConnection;
 
 import org.graalvm.nativeimage.hosted.Feature;
 
@@ -38,6 +42,8 @@ import com.oracle.graal.pointsto.constraints.UnsupportedFeatureException;
 import com.oracle.svm.core.SubstrateOptions;
 import com.oracle.svm.core.annotate.AutomaticFeature;
 import com.oracle.svm.core.image.DisallowedImageHeapObjects;
+import com.oracle.svm.core.jdk.management.ManagementFeature;
+import com.oracle.svm.core.jdk.management.ManagementSupport;
 import com.oracle.svm.core.option.SubstrateOptionsParser;
 import com.oracle.svm.hosted.FeatureImpl;
 import com.oracle.svm.hosted.classinitialization.ClassInitializationFeature;
@@ -54,6 +60,15 @@ public class DisallowedImageHeapObjectFeature implements Feature {
 
     private String[] disallowedSubstrings;
     private Map<byte[], Charset> disallowedByteSubstrings;
+
+    @Override
+    public List<Class<? extends Feature>> getRequiredFeatures() {
+        /*
+         * Ensure that object replaced registered by ManagementFeature runs before our object
+         * replacer.
+         */
+        return Arrays.asList(ManagementFeature.class);
+    }
 
     @Override
     public void duringSetup(DuringSetupAccess access) {
@@ -91,6 +106,8 @@ public class DisallowedImageHeapObjectFeature implements Feature {
             return ((ImageGeneratorThreadMarker) original).asTerminated();
         }
 
+        checkDisallowedMBeanObjects(original);
+
         if (original instanceof String && disallowedSubstrings != null) {
             String string = (String) original;
             for (String disallowedSubstring : disallowedSubstrings) {
@@ -121,6 +138,20 @@ public class DisallowedImageHeapObjectFeature implements Feature {
 
         DisallowedImageHeapObjects.check(original, this::error);
         return original;
+    }
+
+    /** See {@link ManagementSupport} for details why these objects are not allowed. */
+    private void checkDisallowedMBeanObjects(Object original) {
+        if (original instanceof MBeanServerConnection) {
+            throw error("Detected a MBean server in the image heap. This is currently not supported, but could be changed in the future. " +
+                            "Management beans are registered in many global caches that would need to be cleared and properly re-built at image build time.",
+                            original, "Try to avoid initializing the class that stores a MBean server or a MBean in a static field");
+
+        } else if (original instanceof PlatformManagedObject && !ManagementSupport.getSingleton().isRegisteredPlatformManagedObject((PlatformManagedObject) original)) {
+            throw error("Detected a PlatformManagedObject (a MXBean defined by the virtual machine) in the image heap. " +
+                            "This bean is introspecting the VM that runs the image builder, i.e., a VM instance that is no longer available at image run time.",
+                            original, "Try to avoid initializing the class that stores the object in a static field");
+        }
     }
 
     private RuntimeException error(String msg, Object obj, String initializerAction) {
